divide large client class into connection & client
